+++ projects/ngx-testing-tools/src/tests/units/common/rx/rx-box.spec.ts
@@ -1,5 +1,5 @@
 import { BehaviorSubject, ReplaySubject, Subject, Subscription } from 'rxjs';
-import { RxBox } from '../../../../lib/common/test-bed/http/rx-box';
+import { RxBox } from '../../../../lib/common/test-bed/rx/rx-box';
 import { validateArray } from '../../../fixtures/helpers/validators/validate-array';
 
 describe('RxBox', () => {
